+++ src/services/@l13/platforms.ts
@@ -1,22 +1,46 @@
 //	Imports ____________________________________________________________________
 
-
+import type { Platform } from '../../@types/platforms';
 
 //	Variables __________________________________________________________________
 
-export let isMacOs = false;
-export let isWindows = false;
-export let isLinux = false;
+
 
 //	Initialize _________________________________________________________________
 
-if (process.platform === 'darwin') isMacOs = true;
-else if (process.platform === 'win32') isWindows = true;
-else isLinux = true;
+
 
 //	Exports ____________________________________________________________________
 
+export const platform: Platform = detectPlatform();
 
+export let isMacOs = platform === 'macOS';
+
+export let isWindows = platform === 'Windows';
+
+export let isLinux = platform === 'Linux';
+
+export function setPlatform (value: Platform) { // Just for testing
+	
+	isMacOs = value === 'macOS';
+	isWindows = value === 'Windows';
+	isLinux = value === 'Linux';
+	
+}
+
+export function restoreDefaultPlatform () { // Just for testing
+	
+	setPlatform(platform);
+	
+}
 
 //	Functions __________________________________________________________________
 
+function detectPlatform () {
+	
+	if (process.platform === 'darwin') return 'macOS';
+	if (process.platform === 'win32') return 'Windows';
+	
+	return 'Linux';
+	
+}